extract promo cards
