+++ features/productHub/views/ProductHubView.tsx
@@ -19,7 +19,7 @@ import { WithLoadingIndicator } from 'helpers/AppSpinner'
 import { LendingProtocol } from 'lendingProtocols'
 import { useTranslation } from 'next-i18next'
 import React, { FC, Fragment, ReactNode, useMemo, useState } from 'react'
-import { Box, Flex } from 'theme-ui'
+import { Box, Button, Flex } from 'theme-ui'
 
 interface ProductHubViewProps {
   initialNetwork?: ProductHubSupportedNetworks[]
@@ -45,7 +45,7 @@ export const ProductHubView: FC<ProductHubViewProps> = ({
   limitRows,
 }) => {
   const { t } = useTranslation()
-  const { data } = useProductHubData({
+  const { data, refetch } = useProductHubData({
     protocols: [
       LendingProtocol.Ajna,
       LendingProtocol.AaveV2,
@@ -70,6 +70,7 @@ export const ProductHubView: FC<ProductHubViewProps> = ({
 
   return (
     <Fragment key={product}>
+      <Button onClick={refetch}>Refetch</Button>
       <Box
         id="product-hub"
         sx={{
