+++ handlers/product-hub/promo-cards/ajnaLpHandler.ts
@@ -1,0 +1,63 @@
+import {
+  ProductHubItem,
+  ProductHubProductType,
+  ProductHubPromoCards,
+} from 'features/productHub/types'
+import {
+  getAjnaPromoCards,
+  promoCardHowToUseBorrowOnAjna,
+  promoCardsWhatAreAjnaRewards,
+  promoCardWhatIsEarnOnAjna,
+} from 'handlers/product-hub/promo-cards/collections'
+
+export default function (table: ProductHubItem[]): ProductHubPromoCards {
+  const {
+    promoCardETHDAIAjnaBorrow,
+    promoCardETHUSDCAjnaBorrow,
+    promoCardUSDCETHAjnaBorrow,
+    promoCardUSDCWBTCAjnaBorrow,
+    promoCardWBTCUSDCAjnaBorrow,
+    promoCardETHUSDCAjnaEarn,
+    promoCardUSDCETHAjnaEarn,
+    promoCardUSDCWBTCAjnaEarn,
+    promoCardWBTCDAIAjnaEarn,
+    promoCardWBTCUSDCAjnaEarn,
+    promoCardWSTETHDAIAjnaEarn,
+  } = getAjnaPromoCards(table)
+
+  return {
+    [ProductHubProductType.Borrow]: {
+      default: [
+        promoCardETHUSDCAjnaBorrow,
+        promoCardWBTCUSDCAjnaBorrow,
+        promoCardsWhatAreAjnaRewards,
+      ],
+      tokens: {
+        ETH: [promoCardETHUSDCAjnaBorrow, promoCardETHDAIAjnaBorrow, promoCardsWhatAreAjnaRewards],
+        WBTC: [
+          promoCardWBTCUSDCAjnaBorrow,
+          promoCardHowToUseBorrowOnAjna,
+          promoCardsWhatAreAjnaRewards,
+        ],
+        USDC: [
+          promoCardUSDCETHAjnaBorrow,
+          promoCardUSDCWBTCAjnaBorrow,
+          promoCardsWhatAreAjnaRewards,
+        ],
+      },
+    },
+    [ProductHubProductType.Multiply]: {
+      default: [],
+      tokens: {},
+    },
+    [ProductHubProductType.Earn]: {
+      default: [promoCardETHUSDCAjnaEarn, promoCardWSTETHDAIAjnaEarn, promoCardsWhatAreAjnaRewards],
+      tokens: {
+        ETH: [promoCardUSDCETHAjnaEarn, promoCardWhatIsEarnOnAjna, promoCardsWhatAreAjnaRewards],
+        WBTC: [promoCardUSDCWBTCAjnaEarn, promoCardWhatIsEarnOnAjna, promoCardsWhatAreAjnaRewards],
+        USDC: [promoCardETHUSDCAjnaEarn, promoCardWBTCUSDCAjnaEarn, promoCardsWhatAreAjnaRewards],
+        DAI: [promoCardWSTETHDAIAjnaEarn, promoCardWBTCDAIAjnaEarn, promoCardsWhatAreAjnaRewards],
+      },
+    },
+  }
+}
