+++ handlers/product-hub/promo-cards/collections/aaveV2.ts
@@ -1,0 +1,36 @@
+import { NetworkNames } from 'blockchain/networks'
+import { ProductHubItem, ProductHubProductType } from 'features/productHub/types'
+import {
+  findByTokenPair,
+  getEnterWithToken,
+  parseEarnYieldLoopPromoCard,
+} from 'handlers/product-hub/helpers'
+import { LendingProtocol } from 'lendingProtocols'
+
+export function getAaveV2PromoCards(table: ProductHubItem[]) {
+  const aaveV2EthereumProducts = table.filter(
+    ({ network, protocol }) =>
+      protocol === LendingProtocol.AaveV2 && network === NetworkNames.ethereumMainnet,
+  )
+  const aaveV2EthereumEarnProducts = aaveV2EthereumProducts.filter(({ product }) =>
+    product.includes(ProductHubProductType.Earn),
+  )
+
+  const STETHETHAaveV2EthereumEarnProduct = findByTokenPair(aaveV2EthereumEarnProducts, [
+    'STETH',
+    'ETH',
+  ])
+
+  const promoCardSTETHUSDCAaveV2Earn = parseEarnYieldLoopPromoCard({
+    collateralToken: 'STETH',
+    debtToken: 'ETH',
+    pills: [getEnterWithToken('ETH')],
+    product: STETHETHAaveV2EthereumEarnProduct,
+    protocol: LendingProtocol.AaveV2,
+    withYieldExposurePillPill: true,
+  })
+
+  return {
+    promoCardSTETHUSDCAaveV2Earn,
+  }
+}
