+++ handlers/product-hub/promo-cards/collections/index.ts
@@ -1,0 +1,5 @@
+export * from './aaveV2'
+export * from './aaveV3'
+export * from './ajna'
+export * from './generic'
+export * from './maker'
