+++ handlers/product-hub/promo-cards/index.ts
@@ -1,6 +1,6 @@
 import { ProductHubItem, ProductHubPromoCards } from 'features/productHub/types'
-import ajnaLpHandler from 'handlers/product-hub/promo-card-collections-parsers/ajnaLpHandler'
-import home from 'handlers/product-hub/promo-card-collections-parsers/home'
+import ajnaLpHandler from 'handlers/product-hub/promo-cards/ajnaLpHandler'
+import home from 'handlers/product-hub/promo-cards/home'
 import { PromoCardsCollection } from 'handlers/product-hub/types'
 
 export const PROMO_CARD_COLLECTIONS_PARSERS: {
